AhoCorasick.java
This class contains methods to add and remove a string from a set and runs AhoCorasick algorithm to count the occurances of each element in the set in the given string. I spent a lot of time to learn and implement this algorithm. Yet I couldn't solve the problem "String Set Queries" for which I learned this algorithm. I'm getting TLE in the 19th test case.