AhoCorasick using Trie
This class provides methods to add a string to trie, remove a string from the trie and run AhoCorasick algorithm to count the occurences of each string in the set in the given String pattern. It took nearly 3 days for me to learn this algorithm and implement it. Yet I couldn't solve the problem "String Set Queries" from codeforces educational round 16 for which I tried learning this algorithm