Euler's Cycle On Graph
This is a problem from Codeforces. This problem is very simple one and it's the application of "Euler's Cycle" in a graph. Euler's cycle on a graph is the path in a graph which visits each edge only once , and ends at the node from where we started.  From this problem, I learn that Euler's cycle on a graph is possible if and only if each vertex has even degree in graph. To achieve that I added a dummy node 0 and added an edge to 0 to make the degree even. Then I just performed Euler's tour. While taking Euler's tour using DFS, Keep deleting edges , so that those edges won't be visited again but the nodes can be visited.