Lazy propagation , Non uniform updation
This problem is from October challenge 2016. This is one of the most difficult problems that I've ever solved. This is an application of lazy propagation. I had to create new methods for non uniform updation. Implementation was very very difficult and I had a really tough time debugging the application. I stayed in the office till 12:30 to solve this problem.