Sum of Fibonacci Values of Subarrays Sum
This is a nice problem that I solved from hackerrank. I could figure out the logic , I kept getting TLE for a long time. After some time I came to know that the method that I've been using to find the nth fibonacci number is not fast enough. I solved it for 100 points.