Line Sweep Algo, Number of Elements in a Range
This is a nice problem from codechef OCT16 challenge. This is a simple application of 'Sweep line algorithm'. I had to solve a sub problem, find the count of elements in a range. I think it is a standard problem with Binary search tree with AVL rotations. But I found a better and simpler method using Tries. I had to try and debug a lot to get 100 points. I made it :-)