Bitmasking, DP
This is a beautiful problem on the application of bitmasking and dp. I went to office on diwali evening and tried to solve this problem. Then I could'nt understand this editorial. After solving a bitmask dp problem, I got some idea and Implemented on my own. Got AC in the first attempt. I loved this problem.