Centroid Decomposition Of Tree
This is my first algorithm in 2017. This is a technique to decompose the tree which will come in handy in many problems. I've been trying to implement the algo for a long long time. Atlast I successfully implemented and this is faster than the native one, since I cut down the complexity of removing the edges by maintaining another array. Hope to learn a lot in this year. All the best reader :-)